build versioning and functional terminal

+++ inc/kstring.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void* memset(void* bufptr, uint64_t value, size_t size);
+void* memset(void* bufptr, uint32_t value, size_t size);
 void* memcpy(void* restrict dstptr, const void* restrict srcptr, size_t size);
 size_t str_len(const char* s);
 int str_cmp(const char* a, const char* b, size_t n);
